backend2/pyproject.toml
[tool.poetry]
name = "radiology-scheduler"
version = "0.1.0"
description = "Radiology Shift Scheduling System"
authors = ["Rohit Takhar <rohit.takhar@voxirad.ai>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
mysql-connector-python = "^8.2.0"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.11.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-multipart = "^0.0.6"
pydantic-settings = "^2.8.1"
razorpay = "^1.4.2"
setuptools = "^78.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 